Bug fixes - Fixed issue #12 - Make sure all HashMap and SparseArray accesses are thread safe - Removed constants and replaced with #defines for better readability

+++ LiquidCoreAndroid/src/main/java/org/liquidplayer/node/FIleSystem.java
@@ -197,7 +197,8 @@ class FileSystem extends JSObject {
     }
     static void deleteRecursive(File fileOrDirectory) {
         try {
-            if (fileOrDirectory.isDirectory() && !isSymlink(fileOrDirectory)) {
+            if (fileOrDirectory.isDirectory() && !isSymlink(fileOrDirectory) &&
+                    fileOrDirectory.listFiles() != null) {
                 for (File child : fileOrDirectory.listFiles()) {
                     deleteRecursive(child);
                 }
